+++ backend/cwa_lib/validators/messages.py
@@ -1,0 +1,4 @@
+msg_no_special_symbols = """
+Please enter at least 3 characters using only letters, numbers, or common symbols like hyphens (-) and underscores (_).
+No special characters like @, #, $, %, etc.
+"""